requirements.txt
streamlit==1.32.0
boto3==1.34.34
python-dotenv==1.0.1
pymysql==1.1.1